requirements.txt
apache-airflow==2.7.3
scikit-learn==1.3.0
matplotlib==3.7.2
pandas==2.0.3
apache-airflow-providers-amazon==8.11.0